+++ src/AbstractIntId.php
@@ -1,0 +1,31 @@
+<?php
+
+declare(strict_types=1);
+
+namespace CyrilVerloop\DoctrineEntities;
+
+use Doctrine\ORM\Mapping as ORM;
+
+/**
+ * A mapped superclass with an int
+ * as identifier/primary key.
+ *
+ * @ORM\MappedSuperclass
+ */
+#[ORM\MappedSuperclass]
+abstract class AbstractIntId implements IntIdInterface
+{
+    // Traits :
+    use IntId;
+
+
+    // Magic methods :
+
+    /**
+     * The constructor.
+     */
+    public function __construct()
+    {
+        $this->id = null;
+    }
+}