+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.1.0] - 2024-06-19
+### Added
+- `AbstractIntId` and `IntIdInterface`.
+
 ## [8.0.0] - 2024-06-14
 ### Changed
 - `IntId` is now a trait and not an abstract class.